EditContext: add underlineThickness and textColor to TextFormatUpdate event
When TextFormatUpdate event was first introduced [1], ImeTextSpan was
incomplete and we had to map underline thickness to underline style
(which was incorrect). Now we have an updated ImeTextSpan [2], we can
introduce new attributes, textColor and underlineThickness and
correctly return the format info.

Note textUnderlineStyle is renamed to underlineStyle for clarity, and
textDecorationColor is renamed to suggestionHighlightColor since the
value is mapped from ui::ImeTextSpan::suggestion_highlight_color

[1] https://chromium-review.googlesource.com/c/chromium/src/+/1864318
[2] https://source.chromium.org/chromium/chromium/src/+/master:ui/base/ime/ime_text_span.h;l=18;dlc=65548a4764183b6d99b89678d592d49fe3038da5;bpv=1;bpt=0;drc=59cb62c60b608c5aba49542cd59ea95742f2f305

Change-Id: Ib0169931cad2c153565039370ab89b3d8d0b8d9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2425430
Commit-Queue: Alex Keng <shihken@microsoft.com>
Reviewed-by: Kent Tamura <tkent@chromium.org>
Reviewed-by: Anupam Snigdha <snianu@microsoft.com>
Reviewed-by: Yoshifumi Inoue <yosin@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810491}