[IntentHandling] Not showing browser apps in intent picker.
Currently there are couple of issues related to the preferred apps
handling for browser apps in the intent picker.
1. Setting browser apps as preferred app will set it as preferred app
for all urls. This is inconsistent with the behaviour for other apps.
2. If we set the browser as preferred app or set as always stay in
chrome for the browser, then navigate to a website that has other apps
that can handle the url, should we show the intent picker?
It is reasonable for a browser to not suggest other browsers for
handling http/https links, so this CL stops browser apps from appearing
in the intent picker. Users can still open links in other browsers by
opening them explicitly, matching Android's behaviour.

BUG=853604, 1113124

Change-Id: I38b8f7be0d08becca2604ce6998bf40cab8a9729
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2342331
Commit-Queue: Maggie Cai <mxcai@chromium.org>
Reviewed-by: Dominick Ng <dominickn@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810511}