[iOS][Safety Check] Add Omaha support for one off checks
Safety Check on iOS has an update check, to support this we are
adding the ability to do one off checks to Omaha for iOS.

If there is an ongoing scheduled ping, or one is scheduled during the
one off check, it is taken by the one off check.  Otherwise it will
send its own ping.

The one off check updates all user defaults for update status, and
makes Chrome think that it has just shown an update infobar (if an
update is available) so that users triggering an update check aren't
shown an update infobar soon after being told they are out of date by
the update check.

Adds another user default to track if the device is up to date as of
the last check (scheduled or one off). A device that is unable to
update is considered up to date in this user default.

Bug: 1078782
Change-Id: I2e4a1982dd0b34258c78c85ecf23e2f199630f1c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2398725
Commit-Queue: Sean Harrison <harrisonsean@chromium.org>
Reviewed-by: Sylvain Defresne <sdefresne@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810532}