Disable notification mute test on macOS
This test passes on local machines that have a real display attached to
them but fail on bots without. Disable them for now until we find a way
to run them without needing a display. The tests seem to be happy on
Linux at least so keeping them for coverage.

TBR=peter

Bug: 1132058
Change-Id: I0b54d975957c4bc2ff3186a512f04f8f01135768
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2428899
Commit-Queue: Richard Knoll <knollr@chromium.org>
Reviewed-by: Richard Knoll <knollr@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810540}