Greatly simplify tabstrip available width calculation.
Previously, this calculation used LayoutManager::GetAvailableSize,
which relied on cached information that was cleared during layout.
An extra layout was added to fill this cache, which caused no end
of complexities in tabstrip layout.

This is totally unnecessary, as all of the siblings of TabStrip
are simple fixed-width controls - NTB, grab handle, tab search
button, and scroll controls. A wee bit of arithmetic suffices.

Bug: 1093972
Change-Id: I7758bd80ba37b31d5efeeafd208e093094ef5e17
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2426978
Reviewed-by: Connie Wan <connily@chromium.org>
Reviewed-by: Charlene Yan <cyan@chromium.org>
Commit-Queue: Charlene Yan <cyan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810541}