[Android] Show signed in account in sync promo
Sync promo always shows the first account returned by AccountManagerFacade
which might not be signed in. This cl makes sure that the account that
is shown is the first signed in account.

This cl also refactors some of the tests regrading sync promos. The
behavior of the tests are not changed.

Bug: 1131465
Change-Id: I46f2e607494c68d3b0608543095da4b22471f710
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2426485
Commit-Queue: Tanmoy Mollik <triploblastic@chromium.org>
Reviewed-by: Alice Wang <aliceywang@chromium.org>
Reviewed-by: Marc Treib <treib@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810545}