[Autofill Assistant] Refactor lite script notification for ui-shown
This is done in preparation for an upcoming new UKM where we need to
know not just when the UI was shown, but also when the lite script
started. I opted to rename the existing callback instead of introducing
a new one.

This is a refactoring only and should have no user-facing changes.

Bug: b/169052568
Change-Id: I44d4033e20a8c7dfd4cca96be5a3c36388c8b1f6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2424319
Commit-Queue: Clemens Arbesser <arbesser@google.com>
Reviewed-by: Sandro Maggi <sandromaggi@google.com>
Cr-Commit-Position: refs/heads/master@{#810548}