CrOS OOBE - Skip marketing screen for enterprise
The marketing opt-in screen does not show an option to subscribe
to enterprise and supervised users. Since it does not provide any
value and requires the user to click on 'Get Started' there was a
request to skip it altogether.

Bug: 1128918
Change-Id: Iaf40322824b8d75d738e3f58b9a8cce0316c8349
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2414228
Commit-Queue: Renato Silva <rrsilva@google.com>
Reviewed-by: Roman Sorokin [CET] <rsorokin@chromium.org>
Auto-Submit: Renato Silva <rrsilva@google.com>
Cr-Commit-Position: refs/heads/master@{#810553}