[mathml] Update math-depth implementation to match the specification
Experimental font-size: scriptlevel(...) syntax was implemented in [1].
This CL updates it to match what was agreed with the CSSWG [2] as well
as new WPT tests [3]:
- Split scriptlevel(...) into math-depth: ... and font-size: math.
- Rename auto to auto-add

This CL also updates the MathML UA stylesheet to use the new syntax.

To compute font-size, one need the inherited math-style and the
inherited/computed math-depth. So math-style does not need to be High
priority. Additionally Blink computes the High priority alphabetically,
so math-depth must be handled specially.

[1] https://chromium-review.googlesource.com/c/chromium/src/+/2184131
[2] https://github.com/w3c/csswg-drafts/issues/5389
[3] https://github.com/web-platform-tests/wpt/pull/25694

Bug: 6606, 1023292
Change-Id: I8a5bf245fa6158633261ed167800fc5feb0c3564
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2423843
Commit-Queue: Frédéric Wang <fwang@igalia.com>
Reviewed-by: Rune Lillesveen <futhark@chromium.org>
Reviewed-by: Rob Buis <rbuis@igalia.com>
Cr-Commit-Position: refs/heads/master@{#810610}