Reland "Show in-product help when WebUI tab strip is first shown"
This is a reland of 944aea5fd8b08f174dc77daf60afd58f2b5cc439

The test relied on the toolbar being hidden at the start. This CL
shows IPH bubble on the toolbar, causing a toolbar reveal. While this
is desired behvaior in production, the test needs to disable the IPH
to work correctly.

This reland adds a FeaturePromoControllerViews method to cancel the
current IPH promo and block later ones. The test is changed to call it
at the beginning.

Original change's description:
> Show in-product help when WebUI tab strip is first shown
>
> Before, an IPH bubble was shown when a second tab was opened. With
> this CL it is shown as soon as possible after the WebUI tab strip is
> initialized.
>
> Bug: None
> Change-Id: I2f3e7a7dd04229953af5eefe99e00dbcbbfaa8da
> Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2418566
> Commit-Queue: Dana Fried <dfried@chromium.org>
> Reviewed-by: Dana Fried <dfried@chromium.org>
> Cr-Commit-Position: refs/heads/master@{#810097}

Bug: None
Change-Id: Ic3c06f6c35ce9014c3370abc86c029de5af3bb22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2429400
Commit-Queue: Collin Baker <collinbaker@chromium.org>
Reviewed-by: Dana Fried <dfried@chromium.org>
Reviewed-by: Charlene Yan <cyan@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810680}