cros: Refactor window cycle list to decouple scrolling and focus ring.
Currently the logic that handles the focus ring and scrolling the
window cycle list are coupled together. However, there are cases where
the focus ring should move and the list should not scroll and cases
where the focus ring shouldn't move and the list should scroll.

To handle the aforementioned cases, this CL separates the handling for
scrolling and the focus ring.

Test: Manual
Bug: 1067327
Change-Id: I9fed7362c33dd2c5589b8bb5f52e81269273cad4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2425483
Commit-Queue: Jeremy Chinsen <chinsenj@chromium.org>
Reviewed-by: Xiaoqian Dai <xdai@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810732}