[Multipaste] Add context menu option to show the multipaste menu
This CL adds an option to the context menu (with copy/paste options) to
show the clipboard history menu. The CL does the following work:

(1) It creates an interface called ClipboardHistoryController,
enabling the Chrome side to access the clipboard history menu.

(2) It renames ClipboardHistoryController to
ClipboardHistoryControllerImpl which implements the interface.

(3) Adds the option to the context menu. The execution of
the menu option's command should open the clipboard history menu
through the interface.

Bug: 1118158
Change-Id: Idb383571765233514b91d4f2e5629220ddeee1d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2425192
Commit-Queue: Andrew Xu <andrewxu@chromium.org>
Reviewed-by: David Black <dmblack@google.com>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Avi Drissman <avi@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810736}