Remove trivial use of --isolate-script-test-perf-output flag.
The above flag is currently always passed by the recipe to
isolated_script tests. Nothing uses it for anything substantial,
but I forgot when I tried to remove it in https://crrev.com/c/2427305
that there is one trivial use of it (to reconstruct a command line).

This CL removes that one trivial use, so that I can re-land the
original build-side recipe change and remove the flag.

Bug: 1127545
Change-Id: I11b861be35be43f193be24117073a69906c5a73a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2429866
Reviewed-by: Garrett Beaty <gbeaty@chromium.org>
Reviewed-by: Stephen Martinis <martiniss@chromium.org>
Commit-Queue: Dirk Pranke <dpranke@google.com>
Cr-Commit-Position: refs/heads/master@{#810743}