Remove fast/multicol/border-radius-clipped-layer.html test.
It was failing with LayoutNG block fragmentation due to subpixel
differences. Rounded border painting is very sensitive to what layers
are involved, it seems.

In legacy multicol we establish a layer for the flow thread and perform
clipping and translation to slice the tall single column (internal
representation) into columns (what you see on screen). This is not the
case for NG block fragmentation, where we just paint stuff, without any
additional clipping.

The ref had some absolute positioning and clipping. which appeared to
match the layer situation when legacy multicol was used (but not NG).

We already have plenty of coverage for rounded borders in block
fragmentation in wpt/css/css-break/ , e.g. borders-003.html

Bug: 829028
Change-Id: If5ce0626f004acb0466497e0f1eee06ddeccbd38
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2429009
Reviewed-by: Xianzhu Wang <wangxianzhu@chromium.org>
Reviewed-by: Ian Kilpatrick <ikilpatrick@chromium.org>
Commit-Queue: Ian Kilpatrick <ikilpatrick@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810751}