Create ContextMenuPopulatorFactory
Currently, the ContextMenuPopulator is a per-tab object. This
requires us to pass the same set of params to #buildContextMenu() and
then #onItemSelected(). Also, other methods of the class are called with
the same params multiple times.

This CL creates a ContextMenuPopulatorFactory and makes the
ContextMenuPopulator a per-menu object. This way, we can reduce the
number of params the class methods take, and we don't need to call the
same methods with the same set of params many times.

Change-Id: Iacfab7b35fad60cf191ea84443b456f8b60c43c6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2427137
Commit-Queue: Sinan Sahin <sinansahin@google.com>
Reviewed-by: Ted Choc <tedchoc@chromium.org>
Reviewed-by: Jinsuk Kim <jinsukkim@chromium.org>
Reviewed-by: Theresa  <twellington@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810758}