[omnibox] [rich-autocomplete] Rename vars to inline_autocompletion.
inline_autocomplet(ion|e_text)][_] were both used to refer to the
suffix autocompletion of a match and the omnibox. Likewise,
prefix_autocomplet(ion|e_text)[_] were both used to refer to the prefix
autocompletion.

This CL renames the 'e_text' instances to 'ion' as it's both shorter and
the predominant usage.

This is just a renaming change with no behaviour change.

Bug: 1062446
Change-Id: I979b0fc9a383d319d0dcc62ba833b6ae266b32cc
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2427203
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Reviewed-by: Tommy Li <tommycli@chromium.org>
Commit-Queue: manuk hovanesian <manukh@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810787}