Refactor chrome.automation.EventType
This change merges the enums chrome.automation.EventType and
chrome.automation.GeneratedEventType into one enum.

In the native C++ handlers, the two sources of these events is stringified and
assumed to be camel-cased which is what automation expects in js.

This change also shifts enforcement of the relationshipbetween automation and ax
enums from a presubmit to a compilation time warning. At the warning site, a
message is left for the change author to consider the change being made and
actions to take. This applies to both ax events and ax generated events.

Change-Id: I61152f66b2ccf00ba635b2c28caa71fc0caafe5d
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2424854
Commit-Queue: David Tseng <dtseng@chromium.org>
Reviewed-by: Dominic Mazzoni <dmazzoni@chromium.org>
Reviewed-by: Istiaque Ahmed <lazyboy@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810803}