[ios] Refactor the exposed accessibilityHint
Previously we check if the accessibilityHint is set in accessibilityHint method, which blocks the EG test when printing out the UI hierarchy by calling self.accessibilityHint infinitely. Refactor it to customizedAccessibilityHint and expose it in TableViewInfoButtonCell resolve this and make the logic clear. Because for now only settings UI are using TableViewInfoButtonCell and the accessibilityHint for it is customized.

This CL is followed by EG test for settings UI.

Change-Id: Ic56eb1fdd235bca9da06d076b31ee8c9537c805f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2431725
Reviewed-by: Sergio Collazos <sczs@chromium.org>
Commit-Queue: Tina Wang <tinazwang@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810804}