Implement chip view menu item for app menu
Add chip view to downloads/bookmarks menu items.

This CL introduces ChipViewMenuItemViewBinder as a new CustomViewBinder,
it will have a chip view at the end if the feature is enabled.
This CL also moved AppMenuAdapter#OnClickHandler out of AppMenuAdapter,
and renamed to AppMenuClickHandler, because ChipViewMenuItemViewBinder is
the first CustomViewBinder has two items to be clicked, and the click
handler needs to handle clicking items individually.

Bug:1130663

Change-Id: Ia9c0a0350c6e912ee6519d30a07cba2f74a1f72a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2420256
Reviewed-by: Sinan Sahin <sinansahin@google.com>
Reviewed-by: Theresa  <twellington@chromium.org>
Commit-Queue: Gang Wu <gangwu@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810819}