[CrOS PhoneHub] Add phone status view for edge case UIs.
This CL moves the existing |PhoneStatusView| out of the |PhoneHubView|
and adds it as a separate child view of the bubble, so that it can always
show on top of UIs and display available phone information and the
settings icon.

Misc: remove unnecessary header files.

Screenshot: https://screenshot.googleplex.com/6DXqin6MZPcTVVj.png

BUG=1106937,1126208

Change-Id: I321d0511874cfa95dcbb8276fedf41b14e0f684b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2429390
Reviewed-by: Tim Song <tengs@chromium.org>
Reviewed-by: Kyle Horimoto <khorimoto@chromium.org>
Reviewed-by: Xiyuan Xia <xiyuan@chromium.org>
Reviewed-by: Meilin Wang <meilinw@chromium.org>
Commit-Queue: Meilin Wang <meilinw@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810830}