Android MediaRouter: get FragmentManager from embedder.
ApplicationStatus doesn't work outside of //chrome (crbug.com/470582),
so this code must be deferred to the embedder.

A WebContents parameter is included because it's necessary for WebLayer,
even though Chrome doesn't use it.

The MediaRouteDialogManager interface is also removed as it only has a
single implementor, BaseMediaRouteDialogManager.

Bug: 1057100
Change-Id: Ifab92742162cf64e42c75eacfdac6b014940b3df
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2430586
Commit-Queue: Evan Stade <estade@chromium.org>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810837}