Revert: Remove WalkAllDescendants as a speculative fix for IPC crashes
Reverts CL:2148035 as it seems to relate to some IPC crashes.
The reason is unknown, but this safely reverts the change for now.
See repro steps for crash in issue 1094848. The repro steps work
on Canary but not for own-built versions.

TBR: dmazzoni@chromium.org
Bug: 1131848,1094848
Change-Id: I5d13fb13f7b5e9473bdad1f8f80c11a6ca449df1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2433026
Commit-Queue: Aaron Leventhal <aleventhal@chromium.org>
Reviewed-by: Aaron Leventhal <aleventhal@chromium.org>
Cr-Commit-Position: refs/heads/master@{#810935}